Remove unneeded libs and simplify some code. Big-ish.
Removed unneeded libraries:
  addWheelListener globalInput pixiGraphics three threegui viewport
and moved the external libraries to pixigui/external:
  pixi dat
with their non-min versions and maps too (better for development).

Part of the simplification is to avoid the use of axios, which we
do not need for our use in modern browsers (use fetch instead).

There are changes in the bottle server so it serves things from
subdirs of /static/ too (in particular, the libraries in
/static/external).