Use automatic text size from the ete server
We have a new control, automatic size, that creates a new control
on the fly if deactivated. The new control chooses the size, and
is destroyed when going to auto mode again.

There are other minor changes like the max zoom goes from 100 to 1000,
and some fixes like the text appearing in the right position and
taking approximately the right amount of space -- according to the
ete text rectangle.