ayout_fn -> layout_fns and make it a simple list. TreeStyle keeps all defined layouts too.
Removed setters and getters from layout_fn and made it into a simple
list of functions to apply when drawing the content of a node.

Also added a variable self.layouts to TreeStyle so it can keep track
of the available layouts (and thus we can change them after the tree
has been loaded).