Make server error handling work well with gunicorn
This was a hard one: when running interactively with ./server.py, the
error handling was gracefully made and the error was reported in the
response, but not when running it from gunicorn.

The solution I came up with is our own thing, I have not found anything
like that online.