Read quoted content properly
The reading is more conformant to the grammar that appears in
https://en.wikipedia.org/wiki/Newick_format and
https://evolution.genetics.washington.edu/phylip/newick_doc.html

Double escaped single quotes are correctly processed.

The way we read is more permissive that the newick standard, and
the tests in is_valid() do not take into account corner cases when
there are quotations. We may want to create utilities that verify
newicks with different levels of strictness, and help convert into
a more standard format.

Note that the newicks we write do conform to the standard (there are
no extra spaces and all the characters that need escaping are quoted).