Change read() and write() to loads() and dumps()
This is more like in the json module. Also, I have added functions
load() and dump() that work with file objects, the same as in the
json module too.

I also added tests for them. And removed a hack in the server to
show trees that we did not have easily added to the db before.