Add scripts/instance directory
It will contain the trees.db sqlite database. Also, if it has a config.py file
the flask configuration will be overriden with its contents. In particular, this
is useful to set the SECRET_KEY that is used for the token authentication.

The contents of this directory should never be in the repository.