Allow to zoom independently on x and y
The x and y have independent zooms, that are represented in the gui
and can be changed with the mousewheel (pressing ctrl will zoom only
on x, and pressing alt only on y).

Everything is rescaled, which is often not nice for the texts, but it
is a nice first step.

Also made simplifications to the minimap and some cosmetics including
wider use of async functions (for readability).