Make a beautiful zoom in x and y independently
This changes the way the svgs are represented. We do not use a viewbox
anymore, and instead translate the points and sizes in the client
appropriately.

It also changes the initial sizes that are assigned to the content of
the nodes of the tree. Now the content will have a height of either
some minimum or the vertical size of its childs. The horizontal size
will be its length.