Make all texts look and zoom nicely
We change the meaning of the point x,y in a text: it is now the bottom-left
instead of the top-up, in accordance to what it means in svgs and also
to how we normally want to place them.

Also we return a rectangle where the text should fit, instead of a font size,
and leave it to the client to manage it properly.