Allow for different representations
The drawing is now made by "Drawers", which are different representations
or ways to draw the tree (each node with its content_inline, content_float
and content_align). They are like layouts -- I still do not know what is
a good way to name them.

The basic layout (DrawerSimple) is always used for the minimap.