Convert Tree into a cdef class and store sizes more efficiently
We use cython and cdef class Tree to reduce the memory footprint of the
trees and make some functions faster.

Compared with the previous version, a large tree like the ncbi takes 72% of
what it needed in the previous version, and loads it in 63% of the time.

Since we cannot store the size of the content and the childs later, and
the size of the content was redundant in the way we were drawing, we now
store the size of the childs when creating the tree, and have modified
the drawing functions to take that into account.