gui.js: create the svg programmatically
Instead of using div_tree.innerHTML = ..., we now use js to
create all the elements individually and add them to the
document.

It turns out the speed of rendering is similar in chrome, but
seems faster in firefox.

When rendering from a remote server, it does not seem to affect
the speed, which is very likely limited by the connection speed.

In any case, it is a cleaner way to create the drawing, it gives
more control over when to add which element, and most importantly
it will also allow us to render the tree partially when using
websockets -- which will hopefully make it significantly more
responsive.