Make lines centered, have maximum text size, draw less lines
Make the lines centered like a standard phylogenetic tree.

Have a maximum font size when the font size is automatic.

Draw less lines (from parent to its children, now it just a single line instead of n).

Do not draw lines from parent nodes that are not visible in the current viewport. Do not
draw their noderects either (the ones that allow to select a node and zoom on it).

Fix bug for font size when content size == 0 (it would show a big 0, because 0 was a
special number used for floated content text, but now that special number is -1 or any
negative number).