Fix downloads of images
Both the png and svg looked black, because the noderects that we draw
appear as black instead of transparent. That is because the transparency
comes from the css and is not part of the element.

We may want to do it differently in the future, but for the moment we simply
now clone the svg, remove the noderects and then export the image.

Note that the exported image does not have the same colors or thickness, or
any other attribute controlled by css. This is not the best scenario, but
I have not figured out the best solution to it yet.