Circular representation - first draft
We change the rectangular shapes for "annulus sectors", that start at
a point r,a in polar coordinates, and have a size of dr,da.

The radius r plays the role of x in the rectangular representation, and
the angle a goes from amin to amax (-pi to pi by default) in increments
of dy / (amax - amin), where dy are the corresponding increments in the
rectangular representation.

There is quite a bit of duplicated code, and the interesections of the
viewport and the node asecs (annular sectors) is not implemented. Also,
we only have the skeleton of the tree, but no lengths texts or names.

My plan is to merge parts of the rectangular and circular representations
before fixing the missing parts, since I think it will help have a more
compact and understandable structure.