cosmetics
Renamed childs -> children (I did not like it being such a long name, so
I had chosen "childs" even if it is not syntactically corrent, but it is
more of a burden to keep that uncommon name).

d1 -> bh for the branching height (which defines the position where the
branch starts within the node).

Reordered the tree.pyx functions so they split in two more clearly separate
categories, and removed the to_cladogram function that does not really fit
in here, and we are not even using (and also in the spirit of minimalism).