Separate main js code into gui.js and menu.js
Now menu.js initializes the gui. They are js modules, so importing
and exporting are cleaner (and is in strict mode already).

Also, fixed call in tree.pyx init_from_newick() -- cython class is
not happy with named arguments, it seems.

There are some cosmetic changes in the stroke widths and outline
opacity too, to make the trees look nicer by default.