Let the server control the font size.
Instead of the server returning a rectangle and let the client
adjust the font size to it, make the server decide on the font size.

That means the server has to use exactly the available rectangle
in pixels, taking into account the zoom in x and y, which may be
different. But for the floated content we do not want it to depend
on the zoom_x, so the client has to treat especially the case when
the text corresponds to the draw_content_float() (which now it is
just the text of text_type="name").