Center minimap for circular representation. WIP.
It is a Work In Progress because we still need to locate correctly
the minimap visible rectangle, that represents the current view
of the tree.

Also, fixed the zoom in the minimap for the rectangular representation.
It was not taking into account the borders of the minimap (3 pixels
on each side, which come from the .minimap style in the css). As a
result the minimap was overflowing a bit its space. Now it looks
nicer indeed.