+ DrawerCircLengths and DrawerCircFull
This completes the circular drawers. Now we have the same drawers for the
rectangular representation and the circular one.

The code for all of them is extremely similar. The plan is the following:

Instead of deriving DrawerRect and DrawerCirc from the same class, have
the Drawer class get as an input an object that includes their
functionality. Then, from the gui we can choose the option of representing
in a rectangular or circular way, keeping the "Drawer" to whatever we had
(for example, DrawerLeafNames if we only want to draw the leaf names).