Allow to change the angle min and max in the circular representation
I am not completely satisfied with the way the drawers are initialized
separately. It is all a bit hackyish and I would like to find a better
way. But at least it is nice to have the option ready.

Also, changed a few things in the datgui menu, like the position of the
reset-view and the folders for the top-left corner and the zoom.