Simplify tests
Remove the option to launch the tests interactively. Having to decide
what to print and what would make a good tests both in batch mode and
interactively was a burden that made me write less tests and that was
gaining very little.

Now all the tests are of the automatized kind, and intended to be run
with pytest.

Finally, small fix for the test_tree.py:test_str , which uses new
characters to represent the branches.