tree.pyx: add walk() to traverse the tree like in prepostorder
This walk() function is inspired on os.walk() and on the prepostorder
of the original ete.

It yields every internal node twice, first when descending on the tree
and at last when ascending thru it. It also yields a list of the next
descendants of the current node that will be traversed. That list can
be modified (as the dirs in os.walk()) and that will affect how the
traversal continues. In particular, if it is emptied (descendants[:] = [])
the traversal will skip all the children of the current node.

This way of traversing the tree is very economic and will allow functions
like draw.draw() and draw.get_node_box() to be implemented in a
more concise and (hopefully) readable way.