Implement draw() and get_node_box() with new tree.walk()
Also, get_node_box() accepts a function as a parameter, with
a node as an argument, and it will return the first node for
which it is True.

We can then easily implement searching nodes using their name,
or a regexp on their name, or with a complicated expression
that involves the length and the properties of the node, all
from the client-side.