Mark all the nodes in the tree that match the search.
When searching in the gui, instead of zooming into the first
match, now we highlight all the nodes that match (with a
semitransparent yellow rectangle that can be clicked to zoom
into the node). We also show a list of the positions of the
nodes that match, that can be clicked to zoom into them.

Pretty cool :)