Search node using regex or evaluating node properties.
We add the special prefixes /r and /e to mean a search on nodes
that uses regular expresions to match the node name (for /r) and
limited python expressions using properties of the node as its
name, length, etc (for /e). The last one uses a safer version of
eval() which forbids builtins and any name outside of the explicitly
allowed.

We also limit the number of nodes that can be returned, in case we
launch a search on a huge tree and it matches most of the nodes.