draw.pyx: fix the size of the node length text
We were not considering the actual space, only the angle, so the
text with the branch lengths was appearing very small.

I had made the same mistake at the beginning when drawing the leaf
names, also in the circular representation. I think this calls for
a better unification of the drawing modes, so it is automatically
taken into account.