Pass node id information and use it to select subtrees
We walk the trees in a different way now, and in this way
we can easily compute the id of a node (which is a list nid
such that t[nid] is the node).

We use this node id to send that information to the gui, so when
it draws a nodebox it makes it shift-clickable to select that
subtree.