If opening a url with a circular representation, fix the x/y zoom.
If we opened a url with a tree view in circular mode, if the width
and height did not correspond to the original ones the image would
appear distorted. Now we correct for it so the zoom in x and y are
the same for the circular representation, and the original image
that appears in the url is still visible when opening it in a
browser with a different width/height ratio.