Select full nodes (including children) and show them a bit zoomed-out.
When searching a node we now return its node box, not its content box
(that is, the box that contains the node with its children too).

Also, when we zoom into a box, we do not fit it exactly on the screen,
but instead we leave a border of 10% of the space so we see a bit
the surroundings.