tree.pyx: add tree sort function
This allows us to have a way to uniquely identify the nodes.

After reading a newick, the tree has the branches in the same
order as they appear in the newick text. But after calling
sort(), we have the branches appearing in a (almost) unique way,
and then we can refer to the nodes uniquely (by using their
node_id like 0,1,0,0,2,1 or their position when traversing the
tree in preorder, or any other similar way).

The syntax of the sort function is the same as the one python
uses for lists. The "key" argument is a function that gives
a value for a node that will be compared with the ones from
the other sibling nodes. By default, this comparison sorts
by the number of final leaves reachable from that node, and
if equal, by the total length of the node with all its children,
and if equal, alphabetically by name.