draw.pyx:get_node_at advances node according to its content
This is a minor change without consequences for the moment, but
it may have in the future. When traversing the nodes, we advance
in the "y" according to the size of the content of the node, and
not the full node as before, since we may want to use that size
to represent the node including all floating graphical elements.