server.py: send response gzipped.
It turns out this is trivial to do with this flask module, and it makes
a BIG difference!

I removed all my previous uncommited code related to msgpack, which was
way more involved and had gains of about 50% (not bad) compared with
the 90% that we get here (much better!) And almost for free, plus more
standard and all that.