gui.js: do not redraw if we did not drag with the cursor
If we clicked but did not actually drag, we used to make a
request to the backend for a new drawing. This was not
correct. Also, it caused problems when zooming into a node
with ctrl+click, since there was the redrawing of the zoom
and the redrawing of the dragging (with no dragging really),
and they could show the region zoomed in with the wrong zoom.