Include floating graphic elements on the leaves as part of their nodebox (and more, read below).
we need to be able to compute the boxes of the drawn elements (so we
have a bunch of new functions: drawn_size, get_rect, get_asec,
Drawer.get_box).

The font size of the text now is always the vertical size in tree
coordinates, whereas its horizontal size in-tree will be
  len(text)*(fs/1.5)*zy/zx
(and thus, the texts would look distorted when seen in-tree, but
nicely when seen in the viewport / on the screen).

We change the function get_node_boxes() -> get_nodes() , which now
returns the node_id too (so we can keep it in the frontend and use
it to check for nodes searched and being already shown).