Show search results and keep them in the gui.
Many changes.

We keep track of all the searches that have been made, and add
them to the gui (datgui), so we can change the color of their
nodes, and we can remove them from there too.

We show the node_id of the resulting nodes, and the search text
in the results too.

Also, we show the total number of nodes that are being presented
on the screen. This will also help us find if there are problems
in the backend and we are receiving too many nodes.