Combine collapsed nodes and draw efficiently.
We used to draw all the collapsed nodes. Now, instead, we do as we used
to do with the outline boxes: we accumulate them and only when they
cannot be stacked, we call the draw_collapsed() function.

When draw_collapsed() is called, it knows about the box where all the
contiguous and collapsed nodes reside (self.outline), and also about
all the nodes that are contained in that box (self.collapsed). The
drawer can use that information to create a summary in any way it wants.
We include as an example a summary of the names of the nodes, properly
spaced.