Expand in angle around cursor in circular representation.
We can now zoom in and out of a certain point in the circular
representation by expanding in angle. Something similar to what
we can do in the rectangular representation by zooming individually
on x and y.

The zooming keeps the current cursor position in the same place.

For all that, we allow angles < -pi and > pi, with the
understanding that any nodes that do not intersect the annular
section that goes from -pi to pi are always "outside of the
viewport". Which makes sense intuitivelly and allows for a
very nice zooming effect.