Smooth angle zooming.
We reuse the appropriately named zooming.timeout, and
follow the same idea as when delaying the actual zooming,
except in this case we do not rescale the svg meanwhile.

Note that it may be possible to have at least a good
approximation to the re-scaling in the svg, by using
the transform rotate() and scale() and/or matrix().