Show again the search results dialog if clicked on the menu.
Add a new entry in the search folder in the gui, that shows
the search results dialog for each search.

Also, the search results appear no longer as a toast, meaning we can
actually see the top part (which informs on the total number of nodes
found and if there may be more). And we can click outside to close it.