Do not send graphic elements outside of  -pi < angle < pi
If in circular representation we are going to send a childrenline
or nodeline with an angle which is not visible (because we have
zoomed in in angle), do not send it.

It also makes all the draw_* functions behave the same, by yielding
instead of returning, except for the basic graphic functions
(like draw_line()).

This fixes some weird lines that were appearing in the new zoomed-in
angles circular representation.