Update minimap when appropriate when changing angles, and change ctrl-wheel zoom
We mark the minimap as not up-to-date when the tree structure changes.
Right now it is used only for a change in angle_min and angle_max, but
we could use the same idea elsewhere.

We changed the behavior of ctrl-wheel zoomming so it zooms in y (and
let alt-wheel to zoom in x instead). This is consisten with the way
angular zooming works in circular mode.

And it looks cool.

Also, made some cosmetic changes in the global view variable, to
better reflect their structure in the datgui menu.