Adapt to use graphics shifted from the origin manually.
Instead of using a <g transform="translate(...)"> we now do the
shifts manually in the code.

The main reason is because when the numbers in the translations
(and then the xs and ys) were more or less big (like 19979865),
the graphics did not show. This is a limitation of web browsers
and can be seen for example here:
https://oreillymedia.github.io/Using_SVG/extras/ch08-precision.html

This change turned out to be quite involved. But now we should be
able to visualize huge trees (like the complete ncbi tree).