Walk over tree using an iterator.
This is a cosmetic change, that hopefully makes the code clearer.

When walking a tree, we yield an iterator, that can access the
current node, its node_id, and be set to descend=False to stop
descending into the children of the current node. Which is more
clear and less error-prone than overwriting the node_id as we
were doing before.