+ convert.py to create nicer newicks from weird ones.
Some newicks have instead of the node names, their support. And
they include a lot of useless things in their properties, or
lengths of 1 instead of saying they do not have lengths.

The program convert.py tries to guess which format a newick has,
and converts it into a format that is nicer to show with ete.
The resulting newick is written in a file ending in _converted.tree.