tree.pyx: nodes have a parent now.
Only for the root node, parent==None. Otherwise, it points
to the parent of the node.

This will allow certain operations to be done faster and/or
using less memory, but we also will have to be careful to
update the pointers to the parents if we play around with
the trees.