Update minimap gracefully and whenever necessary.
The minimap was not being updated in certain scenarios (like zooming on
angle while the minimap was not visible, and then making it visible
again).

We now use in a smarter way view.minimap_show and view.minimap_uptodate
to decide when we need to really draw again the minimap. This fixes all
the known issues with the minimap so far.

Also changed name put_command() -> api_put().