Smart zooming: xoom towards the node under the mouse.
When we zoom in, we do it so that the node under the cursor
tends to occupy the full screen. When we zoom out, we do it
as before, going out at a constant factor.

We do it by associating the wheel event to the nodes and using their
boxes size information, rather than using the wheel event on the
full document. We only use it on the full document if we are not
over a node.

Similarly, we use now the contextmenu event on the document if we
are not using it on a node, and thus we can show a nice contextmenu
with the options for the full tree, when no nodes is below the cursor.