Make the zoom and the contextmenu work on searched nodes.
Since we were creating new nodes when doing a search, and
they would appear on top of the other nodes, the events
related to mouse wheel and contextmenu would not work on
them. We have added now event listeners for the search
boxes, which are as close as possible to the events of the
actual nodes.