Include collapsed elements in the nodeboxes of collapsed nodes.
Since we only actually drew the box and contents representing the
collapsed nodes when we tried to draw/append a new one that could
not be stacked, we had no way to know the total size of the collapsed
nodebox at the time when we saved it, only its size without the drawn
elements.

With this change, we update the collapsed node when going back in the
tree (by changing the behavior of update_collapsed() so it accepts
None as parameter). This way, we have all the elements related to
collapsed nodes if there are any, at the right time, and we can use
them to compute the total width of the collapsed node including the
things it draws, like we do for all the others.