Refactor drawing code.
The collapsed/outline logic, the node boxes, the node_dxs distances and the
drawing elements were too intertwined in an ugly draw() function.

We better separate the logic and split it in the first time a node is visited
and the last one. The drawing and updating of the collapsed nodes is also
more isolated.

The boxes are now sent as ["b", ...]  rather than using "r" and "s" for
rectangles and annular sectors, since that is what appears naturally in
the current interchange between the backend and gui.js. This simplifies
a bit the logic.

There has been a lot of testing to make all this work nicely, but it can
probably be improved by handling the logic of the boxes and, especially,
the node_dxs, in a separate way, probably using object member elements
(as with self.collapsed).