Fix drawing of angles close to pi, and cosmetics.
To check if an angular interval is valid, we compare to pi+EPSILON, since
rounding errors from using self.y2a can wrongly make angles look slightly > pi.

The outline and the lists of collapsed, boxes and node_dxs that are used during
the walking of the tree for different purposes, are now kept as member variables
and updated appropriately on the functions called from draw().

flush_outline() now returns the box representing the outline flushed (which we
convert into a drawing with draw_box() later). The circular drawer overloads this
function to clip the box between -pi < angle < pi.

Plus a few other cosmetic changes.