Fix saving nodeboxes in frontend.
Since we changed the code for the boxes that are sent from the
backend (now they start with "b" instead of "r" or "s"), we
had to update the is_nodebox() function in the frontend.

The problem that we had before this update is that the searched
nodes would not appear in the drawing, since it did not find
any boxes.