Upload tree: ready for external users.
To upload a tree you do not need to supply a name (it will
be taken from a hash of the current time if not given).

Adding metadata (name, description) is still possible by
clicking on the corresponding checkbox.

It is easier to log in as guest. There is a check to see if
the user is correctly logged at the beginning.

And in general it looks much better.