Smart zoom out.
When zooming out, we now go smoothly towards having the complete tree on
the screen.

For that, we now store the total tree size at the beginning when loading
a tree, and use it instead of making an api call to /trees/<id>/size
each time. We also update it every time that is appropriate (when changing
the tree or modifying it in an UPDATE call).

We have also simplified a bit the smart zooming code when zooming in, and
reused most of it for the zooming out.