add_tree.py: set to 0 the length of the root if it is not specified.
Unfortunately, many trees come without a length for the root node
when they mean it has no length. This is inconsistent with
branches that have no length, where it can mean some fixed distance
(that we take to be 1).

So, when adding trees, we will first set the root to length=0 if
that is what they seemed to want (root with no name and no length).