gui.js: Fix sharing of position, and cleanup initialization.
When we added reset_position() it broke the sharing of a view,
because after loading the x,y,zx,zy etc from the query string,
it would then reset the position and lose the requested one.

We fix that by setting initially x and y to null, and checking
them when reading the query string and doing the rest of the
initialization.

Now, after setting the values from the query string, the internal
values that are stored in the view, like the tree size and if it
is circular, are set up in their own function. Hopefully this is
more clear and less error-prone.