Standardize trees when loading them - solves the numbers (support) appearing instead of names issue.
We now leave the original newick as it is, which is probably the
most sensible thing to do.

To account for the fact that many newicks include the support
instead of the names, we now go thru the loaded tree and replace
the properties and names from the nodes if the names look like
numbers. Also, if it starts with a root with no name and no length
specified, we force its length to 0, which is what is normally meant
(but is inconsistent with the notation for subtrees!).

That is, now we "standardize" the tree when loading it. That way we
can keep the original newicks but also keep the clean semantics.