server.py: cosmetics - upper case of SQL keywords where appropriate.
Make the SQL keywords uppercase, so they can be more easily differenciated
from the name of the tables and so on.

That is the case for the sql statements that appear in strings in the python
code. For the .sql files, we do not need (nor want) that, since the keywords
are recognized by the IDE and colored/highlighted appropriately already.