Fix height of children line when there are collapsed nodes.
When there were collapsed nodes, the line spanning the children looked broken.

To represent it correctly, we need to know who is going to collapse before,
so we need to do it in postorder.

Now that we do it this way, we do not need to compute the "branch height" that
we used to.

https://tree.taiga.io/project/jordibc-ete-smartview/us/6