Nicer aligned region.
The cursor now tips that you can resize the aligned region by
clicking and dragging.

The cursor turns to text-selection when one presses select text
in the menu.

The border of the aligned region is smaller and looks better.

Finally, the padding of the content of the aligned region is
now 0, instead of the 1% that was before. It actually looks
nicer to have the aligned elements appear immediately when
the aligned region starts.