Move api calls and related to new file api.js, and cleanup.
Things that were shared in upload_tree.js and gui.js are now imported
from api.js. With that, those files have been given quite a bit of
cleanup and minor bug fixes too.

There is now a variable view.allow_notifications that controls if we
can use commands that modify the tree. It is set to true except when
we are viewing one of the example trees.