Simplify logic in upload_tree.js and cosmetics.
We now simplify the hash to a 32 bit one, which is shorter (nicer)
to read and we do not need the 64 bit version anyway (this is a
prototype for a small (< 1M) amount of trees uploaded without name).

We also use a nicer example tree that is loaded in the upload page
when clicking on the "load an example" link. Nicer on how it looks
on the text area and also on how it shows in the tree explorer.

All the fetch() operations are now done from api.js , which now has
all the appropriate functions to make api calls.