draw.pyx: Put branching height to half node for non-visible nodes.
This makes the tree change less when the non-visible nodes become
visible as the user moves around the tree. It is the best compromise.

Also, change the name of node_bdy_dys -> bdy_dys. The name is still a
mouthful, but at least it does not say anything about "node".

When I wrote it first, I had in mind self.node_dxs. But in that case,
it is about the dxs of the nodes, as opposed to the dxs of the contents.
In the case of bdy_dys that makes no sense.