Make the dotted line in the aligned texts cover the full viewport.
In DrawerAlign we now draw a line of points from the leaves to the aligned panel
which is as long as needed, and no more.

This is an improvement over the 2*self.tree_dx kludge that I was doing before,
wasting no space and assuring that it will always look nice arriving to the
texts.