Fast drawing and zooming.
When zooming repeatedly, what was taking more time was not the
computations from the backend nor the sending the data, but
actually painting the svg in javascript.

We could paint the full svg first and attach it later, which
would make it much faster, but then we could not use the
getBBox() function that we needed to flip the texts that were
upside down (since it needs the element to be rendered for it
to work). So we were rendering the svg elements one by one as
they arrived (adding them to the g object that was alread
attached to the DOM).

Now, we first draw everything outside of the DOM, and after
adding it, we take the texts that are upside-down and flip them.
This gets us the best scenario: fast and texts looking good.

With this change, we can reduce the default latency for
zooming. We change it from 400ms to 50ms. Since the
redrawings are much faster in the corner cases, the full
experience is now much smoother.