Fast circular text update by prioritizing upside-down texts.
After drawing, the gui could be less responsive for seconds when there
were many upside-down texts to update.

With this change, in addition to the increased refresh speed from the
previous commit, we get a significant boost when there are many texts
upside-down.

We now take all the texts that need to be drawn upside-down, and flip
very well (with BBox) the biggest 1000, and flip only approximately the
rest (which can easily be more than 10k for high resolutions).

The result is indistinguishable (to me) from doing all perfectly, and
it is much faster when there are many such texts. (From ~10s to ~1s).