Show less drawers in the gui and use consistent Rect/Circ names.
Most of the drawers are used as examples and building blocks
for the full rectangular and circular drawers.

We were showing them on the gui nevertheless to experiment with
them. But it does not make much sense now and they distract
from the real functionality that the viewer offers. So, better
to simplify and show only the interesting ones.

Also, use more consistent names (like DrawerRect...).