Server-side searches.
When we search, it is now the backend that keeps the search
results. Thus, when drawing the tree, it will check if
the node is the result of a search and send that information
included in the graphics too.

We do this for both the search results and all their parents.
This way we can send a text that describes the search in the
boxes for the results, and in the lines for the parents. We
use that information to add a class to the appropriate objects
(lines or boxes), and thus we can control their appearance
by changing their style in the frontend.