Remove unused functions - Drawer:get_node_at
These functions, together with the old Drawer:get_nodes(), could
be useful, but we are not currently using them, and it seems
unlikely that we will need them.

The reason is that by using the concept of "node_id" and adding
that information in all the nodeboxes that are sent to draw, we
have an easier way to access a given node on the screen. The
frontend just needs to get that information from the box, and send
it to the backend, which then can access immediately to the node
(the access is very fast).

So there seems to be no incentive to use get_node_at(). And for
the moment I think it is better to remove it to reduce the cognitive
load of the drawing module.