Put back together newick.pyx and tree.pyx, and fix tests.
This commit partially reverts d4c3f1a2fd2.

When using the modules without compiling them with cython, they
would cause an error due to a circular import.

The way trees represent their contents is very tied now to how
they are represented as a newick, so it is not easy to resolve
that circular dependency. I do not think it is really a problem,
though.

In any case, newick.pyx is back into tree.pyx.

And all the tests have been fixed to reflect that and the previous
last changes to the drawing with searches.