Faster check for intersection with viewport in circular representation.
This change speeds up a lot the drawing in the circular representation
when we are seeing a big part of the tree, which happens very often.

Checking that the surrounding rectangle of an element intersects the
viewport should be enough to see if it could be visible. But the
function that gives the surrounding rectangle of an annular sector
(circumrect() is a bad approximation for annular sectors with big
angles (da), so we used to check whether that approximation or the
complementary (and also approximate) one was true too. That part was:
  intersects(self.circumasec_viewport, box)

The problem is that it was true for many elements not visible. And
that was a big slow-down of the program.

By restricting the circumrect() check for small enough angles (<pi/4)
the approximation is good enough, and the results are much better.