Fix fix_text_orientations() so it acts on the texts being drawn.
It was taking the texts always from div_tree, which is not correct,
it should take it from the div that is being drawn. In particular,
when we draw the minimap, it should not act on the texts of div_tree.

This was causing an exception when changing from rectangular view
to circular view.