Tag nodes.
A new option in the context menu for nodes allows to tag the
nodes.

The tags are kept in the frontend, and appear in the menu like
the searches do (you can change their color and opacity, and
remove them too).

You can tag different nodes with the same tag and they will appear
together in the tags menu, and with the same color, etc.

To keep track of the nodes and their svg nodeboxes, we now add
the node_id as id for the svg representing the node.