Simplify how to get the number of visible nodes.
We do not need the save_nodeboxes() function, we can just count
the number of svgs of class "node". And we were not using
view.nodes.boxes anymore (which we used in the past for searches,
when they were done in the frontend -- now they are in the
backend). So we can eliminate that too.