Fix align heatmap and simplify nodeboxes id.
Instead of inserting an id into the svg nodes when constructing
them, we construct the svg elements first, and if they are nodes
we add them the id afterwards.

That way we can still use rectangles (and asecs) for other uses
than nodeboxes. And indeed we were using them for arrays, and
the previous change (which came with the tagging), broke them.

Also, we now use a more friendly id which does not include commas
and does include a prefix of "node-" (so it looks like
"node-0_0_1_0" for example).