Make circumrect() always valid.
The function circumrect(), used to find the horizontal rectangle
that surrounds a given annular sector, was only an approximation
valid for small angles of the annular sector.

By considering the cases where the annular sector crosses the
x/y axis, we can make it exact.

This change makes it exact, and also allows us to simply use it
when we are checking if a box that is an annular sector (in the
circular representation) is in the viewport.

Simpler and more precise, and faster.