Send texts with an associated area (box) where they must be drawn.
We used to send from the backend the font size (before scaling) and
the (bottom-left) point where the text had to be drawn. The frontend
then scaled and drew the text.

When the frontend wanted to limit the size of the texts, they would
always shrink towards the starting (bottom-left) point. Depending
on how we want to draw, this is not what we want (for example, if
we draw the support below the branch line).

With this change, the backend sends the box where the text must go.
The frontend then computes the font size, depending on whether there
is enough width and height for it (or dr and r*da in circular repr.).
When we want a text to be limited only by its vertical size (like
the node names in the leaves), we now use a new function,
dx_fitting_texts, that will assure enough dx in tree coordinates, so
in the frontend the limiting factor will be the dy.

We also send now an anchor point, which is a pair (ax,ay) of relative
coordinates within the box. That is, ax=0 is the left of the box, ax=1
is the right, and any other value is also possible. Thus, (0.5, 0.5)
would anchor the box to its center (the text would shrink towards it
if we limit its size in the frontend).

This is both more general and simplifies the code with respect to
the previous solution.